+++ inspectit-gepard-agent/src/main/java/rocks/inspectit/gepard/agent/instrumentation/hook/configuration/model/MethodHookConfiguration.java
@@ -1,12 +1,12 @@
 /* (C) 2024 */
-package rocks.inspectit.gepard.agent.instrumentation.hook.configuration;
+package rocks.inspectit.gepard.agent.instrumentation.hook.configuration.model;
 
 import java.util.Objects;
 import rocks.inspectit.gepard.config.model.instrumentation.rules.RuleTracingConfiguration;
 
 /**
- * Configuration for one specific method hook. Currently just for tracing. Later we will add
- * entry-/exit-actions and metrics.
+ * Configuration for one specific method hook, which is used to update hooks. Currently just for
+ * tracing. Later we will add entry-/exit-actions and metrics.
  */
 public class MethodHookConfiguration {
