+++ inspectit-gepard-agent/src/main/java/rocks/inspectit/gepard/agent/instrumentation/hook/util/MethodHookFactory.java
@@ -3,7 +3,7 @@ package rocks.inspectit.gepard.agent.instrumentation.hook.util;
 
 import rocks.inspectit.gepard.agent.instrumentation.hook.MethodHook;
 import rocks.inspectit.gepard.agent.instrumentation.hook.action.SpanAction;
-import rocks.inspectit.gepard.agent.instrumentation.hook.configuration.MethodHookConfiguration;
+import rocks.inspectit.gepard.agent.instrumentation.hook.configuration.model.MethodHookConfiguration;
 
 /** Creates method hook objects */
 public class MethodHookFactory {
