+++ inspectit-gepard-agent/src/test/java/rocks/inspectit/gepard/agent/instrumentation/hook/configuration/model/ClassHookConfigurationTest.java
@@ -1,5 +1,5 @@
 /* (C) 2024 */
-package rocks.inspectit.gepard.agent.instrumentation.hook.configuration;
+package rocks.inspectit.gepard.agent.instrumentation.hook.configuration.model;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.when;
